requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile ./requirements.in
#
aiofiles==25.1.0
    # via quart
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.3
    # via -r requirements.in
aiosignal==1.4.0
    # via aiohttp
aiosqlite==0.22.1
    # via -r requirements.in
alembic==1.18.3
    # via -r requirements.in
apscheduler==3.11.2
    # via -r requirements.in
argon2-cffi==25.1.0
    # via -r requirements.in
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
beautifulsoup4==4.14.3
    # via -r requirements.in
blinker==1.9.0
    # via
    #   flask
    #   quart
certifi==2026.1.4
    # via requests
cffi==2.0.0
    # via argon2-cffi-bindings
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   m3u-ipytv
    #   quart
flask==3.1.2
    # via
    #   flask-sqlalchemy
    #   quart
flask-sqlalchemy==3.1.1
    # via -r requirements.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.3.1
    # via sqlalchemy
h11==0.16.0
    # via
    #   hypercorn
    #   wsproto
h2==4.3.0
    # via hypercorn
hpack==4.1.0
    # via h2
hypercorn==0.18.0
    # via quart
hyperframe==6.1.0
    # via h2
idna==3.11
    # via
    #   requests
    #   yarl
itsdangerous==2.2.0
    # via
    #   flask
    #   quart
jinja2==3.1.6
    # via
    #   flask
    #   quart
jsonschema==4.26.0
    # via m3u-ipytv
jsonschema-specifications==2025.9.1
    # via jsonschema
m3u-ipytv==1.0.0
    # via -r requirements.in
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   mako
    #   quart
    #   werkzeug
mergedeep==1.3.4
    # via -r requirements.in
multidict==6.7.1
    # via
    #   aiohttp
    #   yarl
priority==2.0.0
    # via hypercorn
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
psutil==7.2.2
    # via -r requirements.in
pycparser==3.0
    # via cffi
pyyaml==6.0.3
    # via -r requirements.in
quart==0.20.0
    # via
    #   -r requirements.in
    #   quart-flask-patch
quart-flask-patch==0.3.0
    # via -r requirements.in
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r requirements.in
    #   m3u-ipytv
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
soupsieve==2.8.3
    # via beautifulsoup4
sqlalchemy==2.0.46
    # via
    #   -r requirements.in
    #   alembic
    #   flask-sqlalchemy
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   alembic
    #   beautifulsoup4
    #   referencing
    #   sqlalchemy
tzlocal==5.3.1
    # via apscheduler
urllib3==2.6.3
    # via requests
werkzeug==3.1.5
    # via
    #   flask
    #   quart
wsproto==1.3.2
    # via hypercorn
yarl==1.22.0
    # via aiohttp
